# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "연말정산 시뮬레이터 API",
    "description": "근로소득 연말정산 예상 결정세액·환급액 계산 (참고용 시뮬레이션)",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/simulations/calculate": {
      "post": {
        "tags": [
          "simulations"
        ],
        "summary": "저장 없이 계산 (실시간 미리보기)",
        "operationId": "calculate_simulation_api_v1_simulations_calculate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulationInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaxResult"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "리소스 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "입력 검증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/simulations": {
      "post": {
        "tags": [
          "simulations"
        ],
        "summary": "Create Simulation",
        "operationId": "create_simulation_api_v1_simulations_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulationCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationRead"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "리소스 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "입력 검증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "simulations"
        ],
        "summary": "List Simulations",
        "operationId": "list_simulations_api_v1_simulations_get",
        "parameters": [
          {
            "name": "tax_year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tax Year"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SimulationSummary"
                  },
                  "title": "Response List Simulations Api V1 Simulations Get"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "리소스 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "입력 검증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/simulations/{simulation_id}": {
      "get": {
        "tags": [
          "simulations"
        ],
        "summary": "Get Simulation",
        "operationId": "get_simulation_api_v1_simulations__simulation_id__get",
        "parameters": [
          {
            "name": "simulation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Simulation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationRead"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "리소스 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "입력 검증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "simulations"
        ],
        "summary": "Update Simulation",
        "operationId": "update_simulation_api_v1_simulations__simulation_id__put",
        "parameters": [
          {
            "name": "simulation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Simulation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulationUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationRead"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "리소스 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "입력 검증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "simulations"
        ],
        "summary": "Delete Simulation",
        "operationId": "delete_simulation_api_v1_simulations__simulation_id__delete",
        "parameters": [
          {
            "name": "simulation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Simulation Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "리소스 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "입력 검증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/global-income/calculate": {
      "post": {
        "tags": [
          "global-income"
        ],
        "summary": "종합소득세 계산 (근로 + 사업 + 기타소득, 저장하지 않음)",
        "operationId": "calculate_global_income_api_v1_global_income_calculate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GlobalIncomeInput-Input"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalIncomeResult"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "리소스 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "입력 검증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/global-income/simulations": {
      "post": {
        "tags": [
          "global-income"
        ],
        "summary": "Create Global Simulation",
        "operationId": "create_global_simulation_api_v1_global_income_simulations_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GlobalIncomeSimulationCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalIncomeSimulationRead"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "리소스 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "입력 검증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "global-income"
        ],
        "summary": "List Global Simulations",
        "operationId": "list_global_simulations_api_v1_global_income_simulations_get",
        "parameters": [
          {
            "name": "tax_year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tax Year"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GlobalIncomeSimulationSummary"
                  },
                  "title": "Response List Global Simulations Api V1 Global Income Simulations Get"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "리소스 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "입력 검증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/global-income/simulations/{sim_id}": {
      "get": {
        "tags": [
          "global-income"
        ],
        "summary": "Get Global Simulation",
        "operationId": "get_global_simulation_api_v1_global_income_simulations__sim_id__get",
        "parameters": [
          {
            "name": "sim_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Sim Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalIncomeSimulationRead"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "리소스 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "입력 검증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "global-income"
        ],
        "summary": "Update Global Simulation",
        "operationId": "update_global_simulation_api_v1_global_income_simulations__sim_id__put",
        "parameters": [
          {
            "name": "sim_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Sim Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GlobalIncomeSimulationUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalIncomeSimulationRead"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "리소스 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "입력 검증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "global-income"
        ],
        "summary": "Delete Global Simulation",
        "operationId": "delete_global_simulation_api_v1_global_income_simulations__sim_id__delete",
        "parameters": [
          {
            "name": "sim_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Sim Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "리소스 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "입력 검증 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreditMethod": {
        "type": "string",
        "enum": [
          "itemized",
          "standard"
        ],
        "title": "CreditMethod"
      },
...
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorResponse"
      },
...
      "GlobalIncomeSimulationCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "input": {
            "$ref": "#/components/schemas/GlobalIncomeInput-Input"
          },
          "source_simulation_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Simulation Id",
            "description": "불러온 연말정산 시뮬레이션 id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "input"
        ],
        "title": "GlobalIncomeSimulationCreate"
      },
...
      "GlobalIncomeSimulationUpdate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "input": {
            "$ref": "#/components/schemas/GlobalIncomeInput-Input"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "input"
        ],
        "title": "GlobalIncomeSimulationUpdate"
      },
...
      "SimulationCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "input": {
            "$ref": "#/components/schemas/SimulationInput"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "input"
        ],
        "title": "SimulationCreate"
      },
...
      "SimulationInput": {
        "properties": {
          "tax_year": {
            "type": "integer",
            "maximum": 2100.0,
            "minimum": 2000.0,
            "title": "Tax Year"
          },
          "income": {
            "$ref": "#/components/schemas/Income"
          },
          "prepaid_tax": {
            "$ref": "#/components/schemas/PrepaidTax",
            "default": {
              "withholding": 0,
              "previous_employer": 0
            }
          },
          "taxpayer": {
            "$ref": "#/components/schemas/Taxpayer"
          },
          "dependents": {
            "items": {
              "$ref": "#/components/schemas/Dependent"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Dependents",
            "default": []
          },
          "deductions": {
            "$ref": "#/components/schemas/IncomeDeductionInput",
            "default": {
              "national_pension": 0,
              "health_insurance": 0,
              "employment_insurance": 0,
              "housing_rent_loan_repayment": 0,
              "long_term_mortgage_interest": 0,
              "housing_subscription": 0,
              "card": {
                "credit": 0,
                "culture": 0,
                "debit_cash": 0,
                "public_transport": 0,
                "sports_facility": 0,
                "traditional_market": 0
              },
              "venture_direct": 0,
              "venture_fund": 0,
              "national_growth_fund": 0
            }
          },
          "credits": {
            "$ref": "#/components/schemas/TaxCreditInput",
            "default": {
              "pension_savings": 0,
              "irp": 0,
              "insurance_general": 0,
              "insurance_disabled": 0,
              "medical": {
                "general": 0,
                "infertility": 0,
                "premature": 0,
                "specific": 0
              },
              "education": [],
              "donations": {
                "general": 0,
                "hometown": 0,
                "hometown_disaster": 0,
                "political": 0,
                "religious": 0,
                "special": 0
              },
              "monthly_rent": 0
            }
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "tax_year",
          "income",
          "taxpayer"
        ],
        "title": "SimulationInput"
      },
      "SimulationRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "tax_year": {
            "type": "integer",
            "title": "Tax Year"
          },
          "schema_version": {
            "type": "integer",
            "title": "Schema Version"
          },
          "determined_tax": {
            "type": "integer",
            "title": "Determined Tax"
          },
          "total_balance_due": {
            "type": "integer",
            "title": "Total Balance Due"
          },
          "source_simulation_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Simulation Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "input": {
            "$ref": "#/components/schemas/SimulationInput"
          },
          "result": {
            "$ref": "#/components/schemas/TaxResult"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "tax_year",
          "schema_version",
          "determined_tax",
          "total_balance_due",
          "source_simulation_id",
          "created_at",
          "updated_at",
          "input",
          "result"
        ],
        "title": "SimulationRead"
      },
      "SimulationSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "tax_year": {
            "type": "integer",
            "title": "Tax Year"
          },
          "schema_version": {
            "type": "integer",
            "title": "Schema Version"
          },
          "determined_tax": {
            "type": "integer",
            "title": "Determined Tax"
          },
          "total_balance_due": {
            "type": "integer",
            "title": "Total Balance Due"
          },
          "source_simulation_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Simulation Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "tax_year",
          "schema_version",
          "determined_tax",
          "total_balance_due",
          "source_simulation_id",
          "created_at",
          "updated_at"
        ],
        "title": "SimulationSummary"
      },
      "SimulationUpdate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "input": {
            "$ref": "#/components/schemas/SimulationInput"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "input"
        ],
        "title": "SimulationUpdate"
      },
...
      "TaxResult": {
        "properties": {
          "tax_year": {
            "type": "integer",
            "title": "Tax Year"
          },
          "rules_verified": {
            "type": "boolean",
            "title": "Rules Verified"
          },
          "annual_earned_income": {
            "type": "integer",
            "title": "Annual Earned Income"
          },
          "non_taxable_income": {
            "type": "integer",
            "title": "Non Taxable Income"
          },
          "gross_salary": {
            "type": "integer",
            "title": "Gross Salary"
          },
          "earned_income_deduction": {
            "$ref": "#/components/schemas/BreakdownItem"
          },
          "earned_income_amount": {
            "type": "integer",
            "title": "Earned Income Amount"
          },
          "applied_method": {
            "$ref": "#/components/schemas/CreditMethod"
          },
          "method_comparison": {
            "items": {
              "$ref": "#/components/schemas/MethodComparison"
            },
            "type": "array",
            "title": "Method Comparison"
          },
          "income_deductions": {
            "items": {
              "$ref": "#/components/schemas/BreakdownItem"
            },
            "type": "array",
            "title": "Income Deductions"
          },
          "total_income_deduction": {
            "type": "integer",
            "title": "Total Income Deduction"
          },
          "tax_base": {
            "type": "integer",
            "title": "Tax Base"
          },
          "tax_rate": {
            "type": "string",
            "title": "Tax Rate"
          },
          "calculated_tax": {
            "type": "integer",
            "title": "Calculated Tax"
          },
          "tax_reduction": {
            "type": "integer",
            "title": "Tax Reduction"
          },
          "tax_credits": {
            "items": {
              "$ref": "#/components/schemas/BreakdownItem"
            },
            "type": "array",
            "title": "Tax Credits"
          },
          "total_tax_credit": {
            "type": "integer",
            "title": "Total Tax Credit"
          },
          "determined_tax": {
            "type": "integer",
            "title": "Determined Tax"
          },
          "prepaid_tax": {
            "type": "integer",
            "title": "Prepaid Tax"
          },
          "balance_due": {
            "type": "integer",
            "title": "Balance Due"
          },
          "local_income_tax": {
            "$ref": "#/components/schemas/LocalIncomeTax"
          },
          "total_balance_due": {
            "type": "integer",
            "title": "Total Balance Due"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/CalcWarning"
            },
            "type": "array",
            "title": "Warnings"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "tax_year",
          "rules_verified",
          "annual_earned_income",
          "non_taxable_income",
          "gross_salary",
          "earned_income_deduction",
          "earned_income_amount",
          "applied_method",
          "method_comparison",
          "income_deductions",
          "total_income_deduction",
          "tax_base",
          "tax_rate",
          "calculated_tax",
          "tax_reduction",
          "tax_credits",
          "total_tax_credit",
          "determined_tax",
          "prepaid_tax",
          "balance_due",
          "local_income_tax",
          "total_balance_due",
          "warnings",
          "disclaimer"
        ],
        "title": "TaxResult"
      },
...
    }
  }
}